requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
django-leaflet==0.28.2
django-mathfilters==1.0.0
django-pdfkit==0.3.1
djangorestframework==3.13.1
EasyProcess==1.1
gunicorn==20.1.0
numerize==0.12
pdfkit==0.6.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
PyVirtualDisplay==2.2
sqlparse==0.4.2
whitenoise==6.0.0
